re-enable mock check on serviceusage

+++ apis/cloudbuild/v1beta1/workerpool_types.go
@@ -49,7 +49,7 @@ type WorkerConfig struct {
 	// +optional
 	MachineType string `json:"machineType,omitempty"`
 	// +optional
-	DiskSizeGb int `json:"diskSizeGb,omitempty"`
+	DiskSizeGb int64 `json:"diskSizeGb,omitempty"`
 }
 
 type NetworkConfig struct {
